% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{addRepasAliment}
\alias{addRepasAliment}
\title{addAliment}
\usage{
addRepasAliment(history, aliment)
}
\arguments{
\item{history}{history object}

\item{aliment}{aliment as data.frame}
}
\description{
addAliment
}
