% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createRepasAliment}
\alias{createRepasAliment}
\title{createAliment}
\usage{
createRepasAliment(name, quantity, repasType)
}
\arguments{
\item{name}{name}

\item{quantity}{quantity}

\item{repasType}{repasType}
}
\description{
createAliment
}
