% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/utils.R
\name{onError}
\alias{onError}
\title{onError}
\usage{
onError(err, functionName, step)

onError(err, functionName, step)
}
\arguments{
\item{err}{err}

\item{functionName}{functionName}

\item{step}{step}

\item{err}{error message in the try catch}

\item{functionName}{name of function the try catch is located}

\item{step}{name of the step in the function}
}
\description{
onError

onError
}
